Send initial, empty dataset to parent (LARA / AP)
[#175787319]
